Refactored svc -> app and corrected mcp mount path issue [ref #13]
