+++ .env.example
@@ -38,5 +38,5 @@ OLLAMA_EMBEDDING_MODEL=embeddinggemma
 OLLAMA_LLM=deepseek-v3.2
 
 ### Vapor Application Service Layer ###
-SVC_DOCKER_HOST_NAME=vapor-svc
-SVC_PORT="8000"
+APP_DOCKER_HOST_NAME=vapor-app
+APP_PORT="8000"
