+++ vapor/app/main.py
@@ -1,0 +1,3 @@
+from .factory import create_app  # pragma: nocover
+
+app = create_app()  # pragma: nocover
